learning_bi/requirements.txt
django-admin-interface
django-ckeditor
